tests/wpt/web-platform-tests/html/canvas/offscreen/the-offscreen-canvas/size.attributes.reflect.setidl.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: size.attributes.reflect.setidl</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>size.attributes.reflect.setidl</h1>
<p class="desc">Setting IDL attributes updates IDL and content attributes</p>


<script>
var t = async_test("Setting IDL attributes updates IDL and content attributes");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

offscreenCanvas.width = 120;
offscreenCanvas.height = 60;
_assertSame(offscreenCanvas.width, 120, "offscreenCanvas.width", "120");
_assertSame(offscreenCanvas.height, 60, "offscreenCanvas.height", "60");
t.done();

});
</script>
